chore: clean codes

+++ spring-graphql-webflux/src/main/java/com/example/demo/repository/AuthorRepository.java
@@ -76,7 +76,7 @@ public class AuthorRepository {
 
     }
 
-    public Mono<Integer> deleteAll() {
+    public Mono<Long> deleteAll() {
         String sql = "DELETE FROM users";
         return this.client.sql(sql).fetch().rowsUpdated();
     }
